Introduce a configurable password grace period
When you're in front of the computer reading something, and the screen lock
blanks the screen, you want to quickly dismiss it without typing the password.

The patch adds an PREAUTH state and a timer that moves to the IDLE state.
If a key is pressed in the PREAUTH state, swaylock just exits.